ProjectFiles/requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
requests
PyJWT
cryptography
replicate
